Code reorg after review comments
